services/analytics-svc/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
pydantic==2.6.3
python-jose[cryptography]==3.3.0
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
